+++ lib/generators/ductwork/install/templates/db/create_ductwork_transitions.rb
@@ -28,5 +28,16 @@ class CreateDuctworkTransitions < Ductwork::Migration
       table.datetime :completed_at
       table.timestamps null: false
     end
+
+    if mysql?
+      add_index :ductwork_transitions,
+                %i[branch_id completed_at started_at],
+                name: "index_ductwork_transitions_on_latest_open"
+    else
+      add_index :ductwork_transitions,
+                %i[branch_id started_at],
+                where: "completed_at IS NULL",
+                name: "index_ductwork_transitions_on_latest_open"
+    end
   end
 end