+++ lib/generators/ductwork/update/templates/db/add_pipeline_started_index_to_ductwork_runs.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddPipelineStartedIndexToDuctworkRuns < Ductwork::Migration
+  def change
+    add_index :ductwork_runs, %i[pipeline_id started_at]
+  end
+end